changing project structure
solving git issue
