+++ app/config.py
@@ -1,0 +1,11 @@
+import os
+from dotenv import load_dotenv
+
+# Carrega as variáveis de ambiente do arquivo .env
+load_dotenv()
+
+UPLOAD_FOLDER = "uploads"
+os.makedirs(UPLOAD_FOLDER, exist_ok=True)
+PDF_CONTENT = ""
+
+OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")
